+++ src/blogs/future.md
@@ -1,5 +1,5 @@
 ---
-title: Futureposting
+title: A Post About Apes From The Future
 date: 2024-09-10
 author: Skeletom
 brief: Hey! Welcome to my brand new developent blog!
@@ -14,3 +14,5 @@ Hey! Welcome to my brand new development blog!
 <!--more-->
 
 This is the second post. Wow!
+
+<iframe width="560" height="315" src="https://www.youtube.com/embed/YiOyPQ8dC1I?si=x_H2po3Ud7HFzvcA" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
