+++ src/templates/home.html
@@ -1,0 +1,95 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title class='meta-title'>skeletom.net</title>
+        <meta class='meta-title' property="og:title" content="skeletom.net">
+        <meta class='meta-desc' name="description" content="Skeletom Dot Net: 'It's Dot Com!'">
+        <meta class='meta-desc' property="og:description" content="Skeletom Dot Net: 'It's Dot Com!'">
+        <meta class='meta-img' property="og:image" content="">
+        <meta class='meta-img' name="twitter:image" content="">
+        <meta property="og:type" content="website">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link rel="icon" type="image/x-icon" href="favicon.ico">
+        <link rel="stylesheet" type="text/css" href='/css/style.css'>
+        <link rel="stylesheet" type="text/css" href='/css/font-awesome.min.css'>
+    </head>
+    <body>
+        <main>
+            <div class="max-width column">
+                <div class="row">     
+                    <div class="white item header">
+                        <span class="title">skeletom.net</span>
+                        <div class="darker item-label-below item">
+                            <span><a href="/">home</a> | <a href="/projects">projects</a> | <a href="/blogs">blogs</a> | <a href="/contact">contact</a></span>
+                        </div>
+                    </div>
+                </div>
+                <div class="row homepage-row">
+                    <div class="dark item div1">
+                        <div class="darker item-label item">
+                            <span class="highlight meta-date">Bone giorno!</span>
+                        </div>
+                        <div>
+                            <h1 class="meta-title">Tom "Skeletom" Farro</h1>
+                        </div>
+                        <!-- <div>
+                            <img class="meta-img thumbnail" src="/img/skeletom_bust.png">
+                        </div> -->
+                        <div id="content">
+                            <img class="meta-img thumbnail float-left" src="/img/skeletom_bust.png">
+                            <p>
+                            <span class="highlight">Ciao!</span>
+                            Like all members of the living dead, I am bound to this world by unfinished business. In my case, it's outstanding <a href="https://www.github.com/fomtarro" target="_blank">Git issues</a>.
+                            </p>
+                            <p>
+                                My stream content includes programming, cooking, model-building, and on rare occasions, even video gaming.
+                                I work day in and day out to make cool tech for both VTubers and their fans, to help everyone enjoy this virtual world to the fullest.
+                            </p>
+                            <!-- blog goes here -->
+                        </div>
+                    </div>
+
+                    <div class="dark item div2">
+                        <div class="darker item-label item">
+                            <span class="highlight meta-date">Stay in touch!</span>
+                        </div>
+                        <div class="contact-column">
+                            <p>Email me at:</p>
+                            <div class="row contact-item">
+                                <span class="fa fa-envelope fa-lg"></span><a href="mailto:tom@skeletom.net">tom@skeletom.net</a>
+                            </div>
+                            <p>Or find me on:</p>
+                            <div class="row contact-item">
+                                <span class="fa fa-twitter fa-lg"></span><a href="https://www.twitter.com/fomtarro" target="_blank">@fomtarro</a>
+                            </div>
+                            <div class="row contact-item">
+                                <span class="fa fa-twitch fa-lg"></span><a href="https://www.twitch.tv/skeletom_ch" target="_blank">skeletom_ch</a>
+                            </div>
+                            <div class="row contact-item">
+                                <span class="fa fa-youtube fa-lg"></span><a href="https://www.youtube.com/@fomtarro" target="_blank">@fomtarro</a>
+                            </div>
+                            <div class="row contact-item">
+                                <span class="fa fa-github fa-lg"></span><a href="https://github.com/FomTarro" target="_blank">fomtarro</a>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <div class="dark item">
+                    <div class="darker item-label item">
+                        <span class="highlight meta-date">Latest Work</span>
+                    </div>
+                    <div>
+                        <h1 class="meta-title"></h1>
+                    </div>
+                    <div>
+                        <img class="meta-img thumbnail">
+                    </div>
+                    <div id="content">
+                        tom@skeletom.net
+                        <!-- blog goes here -->
+                    </div>
+                </div>
+            </div>
+        </main>
+    </body>
+</html>
